week8/virtual_realtor/pyproject.toml
[project]
name = "1-admin-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.206.0",
    "constructs>=10.0.0,<11.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
